Moved movie components from festival to movie lib and created shell
