--- apps/festival/festival/src/app/marketplace/title/view/view.component.html
@@ -1,19 +1,0 @@
-<ng-container *ngIf="movie$ | async as movie">
-  <movie-view [navLinks]="navLinks">
-    <movie-header [movie]="movie">
-      <movie-header-cta fxLayout="column" fxLayoutAlign="start start" fxLayoutGap="32px">
-        <ng-container *ngFor="let org of orgs$ | async">
-          <org-chip [org]="org"></org-chip>
-          <a mat-stroked-button color="primary" [href]="'mailto:' + org.email">
-            <span>Contact Sales Agent</span>
-            <mat-icon svgIcon="arrow_forward"></mat-icon>
-          </a>
-        </ng-container>
-      </movie-header-cta>
-      <movie-header-actions fxLayout.lt-sm="column" fxLayout="row" fxLayoutAlign="start center" fxLayoutGap="24px">
-        <wishlist-button [movieId]="movie.id"></wishlist-button>
-        <movie-promotional-links [movie]="movie" [links]="promoLinks"></movie-promotional-links>
-      </movie-header-actions>
-    </movie-header>
-  </movie-view>
-</ng-container>
