--- libs/movie/src/lib/movie/layout/view/view.component.html
@@ -1,35 +1,0 @@
-<ng-template pageBar targetId="nav">
-  <header fxLayout fxLayoutAlign="start center">
-    <app-menu></app-menu>
-    <ng-container *ngFor="let link of navLinks">
-      <a mat-button [color]="rla.isActive ? 'primary' : ''" [routerLink]="link.path" routerLinkActive
-        #rla="routerLinkActive">
-        {{link.label}}
-      </a>
-    </ng-container>
-  </header>
-</ng-template>
-
-<ng-content select="movie-header"></ng-content>
-
-<nav id="nav" mat-tab-nav-bar>
-  <a mat-tab-link *ngFor="let link of navLinks" [routerLink]="link.path" routerLinkActive #rla="routerLinkActive"
-    [active]="rla.isActive">
-    {{link.label}}
-  </a>
-</nav>
-<router-outlet></router-outlet>
-<movie-screening></movie-screening>
-<h2 flex *ngIf="isEnoughPicturesThen(1) | async">Images</h2>
-<ng-container *ngIf="isEnoughPicturesThen(4) | async; else noCarousel">
-  <bf-carousel flex>
-    <img *ngFor="let img of (movie$ | async).promotional.still_photo | keyvalue" carouselItem [col]="3"
-      [ref]="img.value" asset="empty_poster.webp">
-  </bf-carousel>
-</ng-container>
-<ng-template #noCarousel>
-  <div fxLayout flex>
-    <img *ngFor="let img of (movie$ | async).promotional.still_photo | keyvalue" [ref]="img.value"
-      asset="empty_poster.webp" [col]="3">
-  </div>
-</ng-template>
