+++ libs/movie/src/lib/movie/marketplace/screening/screening.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
 import { EventService } from '@blockframes/event/+state/event.service';
-import { ViewComponent } from '../view/view.component';
+import { MovieShellComponent } from '../shell/shell.component';
 import { Event } from '@blockframes/event/+state';
 import { Observable } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
@@ -16,7 +16,7 @@ export class ScreeningComponent implements OnInit {
   events$: Observable<Event[]>;
 
   constructor(
-    private parent: ViewComponent,
+    private parent: MovieShellComponent,
     private service: EventService,
     private dynTitle: DynamicTitleService,
   ) { }
