+++ libs/movie/src/lib/movie/marketplace/shell/shell.component.scss
